Correctness: exclude words that don't match letter placements exactly
Due to a logic mistake we previous only required that words matched
the letters where there were hits, but not that they _didn't_ match
the guessed letters where there weren't. So, with a hit-list like
['a':1, 't':2], the word "attach" would be considered valid even though
we know that the letters at 3 and 4 aren't 'a' or 't'.

This requires a drastic change to the schema. Previously the wide redundant
representation with all letters in each column of `words` was much faster
due to a much smaller table at least, and possibly for the ease of the
index masking given it was all conditioning the same table. Now we must
resort to self-joins of `letter_agg` since it's the only thing that holds
all the letter locations per row. I'll experiment with a different redundant
representation with one column per letter of the alphabet, which may
compete with the letter_agg self-joins.