Use wide table query on arrays of letter incidences
Instead of a table with columns as the letters in respective character
positions, use a table with columns as arrays of the respective character
positions (e.g. `na` is the array of positions of `a` in the word, etc.)
Although slightly slower, this is much more precise.

Also try to use a GIN-indexed array of letters in the word to try to speed
up the left outer join on the failed set of letters. At least, table scans
of this should be shorter just because there are so many fewer words than
letters.